Use plain sh instead of Bash
The commands used to start the different processes in the Supervisor
configuration in tests work just fine with plain sh, so there is no need to
require Bash.

Similarly, the examples in the readme are simple enough to not require Bash
either, so refer to the former in the Markdown fenced block.