hdf5_to_cat2: Do not force string decoding of bytes attributes
There may be attributes of other types which Blosc2 should be able to
serialize with msgpack into the vlmeta.

The current issue with storing `bytes` attribute values is that NumPy uses
`numpy.bytes_` (a subclass of `bytes`), but Blosc2 uses `strict_types=True`
when serializing, so subclasses are not allowed.

Of course that means that we either need Blosc2 to encode with
`strict_types=False`, or we need to manually narrow values to standard Python
classes first.