Doc: Include readme as content of "Getting Started" section
So as to avoid repetition between them.

Old subsections ("Overview", "Installation") still need to be merged in.

A small kludge in the form of a symbolic link to an image file was needed to
properly include the Caterva2 logo both in the bare readme (e.g. as rendered
by GitHub) and in the Sphinx-rendered docs.