Doc: Include link to `cat2import` doc section from readme
To check how it renders in the plain readme.  The chosen approach is readable
in Markdown source and it renders readable text in all cases (thanks to the
explicit link text) hinting on how to find the target in the full
documentation, although the link produced by GitHub for the readme is of
course broken.

Rule of thumb for such readme links: `(see also [<link text>](<link target>)
in Caterva2 <section> documentation)`.