hdf5_to_cat2: Always set `blocks` argument on Blosc2 node import
Especially for the case of a 1D array where the blockshape was set, since no
chunkshape is stored in filter arguments, so the loaded schunk (not array) may
end up having a different blocksize (and even chunkshape).  For the 1D array,
the blockshape is computed from the blocksize and the typesize.