Fix HDF5 import passing individual smaller Blosc2 chunks
When the Blosc2 super-chunk in the HDF5 chunk consists of many chunks that
would result in too small chunks in the imported Blosc2 dataset.  Such
super-chunks are not supported now, but they should get re-compressed as
full-size chunks.  Anyway this is a corner case (that would need manual
crafting of the HDF5 chunk) and the Blosc2 HDF5 filter currently only produces
super-chunks with one chunk.